+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 The change log itself is written the way that [keepachangelog.com](http://keepachangelog.com/) describes.
 
+## [1.2.0] - 2016-11-29
+### Added
+- dedicated param bag on the request object to store processor related options
+  e.g. spas hooks can flag requests for repetition / simply store other processing related data on them
+
 ## [1.1.0] - 2016-11-29
 ### Added
 - parsed requests know about their respective api resource group objects